tests/drivers/spi/spi_loopback: Fix configurations for nRF boards
Since commit 22326f6d99e8e710bbf7e6f9057bb96741e970f6 got in, this test
requires that so-called NOP frames (bytes of value 0) are sent when no
TX data is provided. For nRF SoCs, the over-read character (ORC) that
is sent in such case is by default configured to the value of 0xFF.
Alter configurations for nRF boards so that 0x00 is used instead.

On the occasion, remove redundant selections of driver type in those
configurations (CONFIG_SPI_x_NRF_SPI[M]=y), as such selections must
be aligned with the value of "compatible" property in corresponding
nodes in dts, so there's no much sense in having them separately here.

Correct also the SPI instance to be used for nrf51_pca10028.

Signed-off-by: Andrzej Głąbek <andrzej.glabek@nordicsemi.no>