Bluetooth: SMP: Fix bond lost on pairing failure.
Fix an an issue where established bonding information in the peripheral
are deleted when the central does not have the bond information.
This could be because the central has removed the bond information, or
this is in fact not the central but someone spoofing it's identity, or
an accidental RPA match.

This is a regression from: a3e89e84a801d9bc048b0ee2177f0fb11d1a925a

Signed-off-by: Joakim Andersson <joakim.andersson@nordicsemi.no>